Remove dependency on num module
The num module drags in a bunch of other submodules. However, the
dependency on num is rather superficial. It is only required to ensure a
zero value for the current type, which is always a number primitive.
However, the existing number primitive types already implement the Default
trait whose default() method returns zero. Switching to this method removes
the dependency on the num module altogether.

Although it is logically possible that default() might be changed in the
future to return a non-zero value, it will never happen in reality. This is
a reasonable trade-off to reduce the number of dependencies.