Expose "the" RawBitFlags trait in the public API
Resolves #23.

This is necessary to make blanket implementations possible. We export
a proxy trait, and keep the real thing in the _internal module.
This "prevents" users from relying on the trait items (either by
implementing the trait manually or calling its methods), while still
being able to name the trait in generic bounds.