Only run UI tests on nightly
One compiler version is enough to detect problems, and the differences
in output between versions are superficial.

We choose to test on nightly as stable compilers do not support the full
span API - only the first token gets highlighted. We want to make sure
that the proper end of the span is chosen.