Add some #[must_use] and #[inline] annotations.
Clippy pointed out the lack of #[must_use] on not_c. It's not like
there's a bit chance of a mistake there, but it doesn't hurt so
whatever.

An #[inline] annotation is necessary for inlining across crates if LTO
is not enabled, so do that to every function which could conceivably end
up in a hot loop.