Run clippy on the test suite
This makes sure that we're not responsible for Clippy triggering in the
user's code.

Not sure whether Clippy is *supposed* to trigger in macro output, but it
is what it is.

To run clippy on the test suite we need to use the --all --all-targets
flags, which have a side effect of Clippy finding the benches/ directory
even though it has been commented out in Cargo.toml. This means we can
no longer keep the criterion dev-dependency commented out.