Make make_bitflags! sound even when dirty tricks are employed
Previously, the macro relied on an expression of the form Enum::Variant
always being a variant of the enum. However, it may also be an
associated integer constant, in which case we don't have any
guarantee that the value only contains bits valid for this enum.

Thus, adverserial input to the macro could create a `BitFlags<Enum>`
with bits that don't correspond to any of the enum variants. Iterating
over such a value would then cause UB. The debug formatter is also
affected, as internally it iterates over the value being formatted.