Expand traits implemented by BitFlags
PartialOrd and Ord are now implemented. The values returned by the
traits aren't meaningful, but they allow using BitFlags as a key
in a BTreeMap. I wish the traits had a distinction between
"meaningful order" (overloaded comparison operators) and "arbitrary
order" (can be used in BTreeMap).

Moreover, the implementation for Hash is improved, so that code
that is generic over BitFlag can still access it.

Closes #47